# expensetrackerapi/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Expense Tracker",
    "description": "Expense Tracker",
    "version": "1.0.0"
  },
  "host": "localhost:1337",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Login service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userCode",
            "in": "header",
            "description": "User code.",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Password.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in successfully."
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Logout service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Token.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User logged out successfully."
          }
        }
      }
    },
    "/validate": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "User Validation service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Token.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User is authenticated."
          }
        }
      }
    },
    "/settoken": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Set Token service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Token.",
            "required": true,
            "type": "string"
          },
          {
            "name": "usercode",
            "in": "header",
            "description": "User code.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Token Successfully Saved."
          }
        }
      }
    },
...
  }
}